requirements.txt
flask
groq
python-dotenv
requests
Pillow
# Optional: inference-sdk (for Roboflow ingredient detection, requires Python <3.14)
